# docs/swagger.yaml
basePath: /api/
definitions:
  handler.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  models.EndRentOutput:
    properties:
      rent:
        $ref: '#/definitions/models.GetRentOutput'
    type: object
...
  models.GetRentOutput:
    properties:
      ID:
        type: integer
      account:
        type: integer
      finalPrice:
        type: number
      priceType:
        type: string
      priceUnit:
        type: number
      timeEnd:
        type: string
      timeStart:
        type: string
      transport:
        type: integer
    type: object
  models.GetRentTransportHistoryOutput:
    properties:
      rents:
        items:
          $ref: '#/definitions/models.GetRentOutput'
        type: array
    type: object
  models.GetRentTransportNewInput:
    properties:
      rentType:
        type: string
    type: object
  models.GetRentTransportNewOutput:
    properties:
      rent:
        $ref: '#/definitions/models.GetRentOutput'
    type: object
  models.GetRentTransportOutput:
    properties:
      transports:
        items:
          $ref: '#/definitions/models.GetTransportOutput'
        type: array
    type: object
  models.GetRentsOutput:
    properties:
      rents:
        items:
          $ref: '#/definitions/models.GetRentOutput'
        type: array
    type: object
  models.GetTransportOutput:
    properties:
      canBeRented:
        type: boolean
      color:
        type: string
      dayPrice:
        type: number
      description:
        type: string
      id:
        type: integer
      identifier:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      minutePrice:
        type: number
      model:
        type: string
      ownerId:
        type: integer
      transportType:
        type: string
    type: object
...
info:
  contact: {}
  description: API for transportation rental service
  title: SimbirGoAPI
  version: 1.0.0
paths:
...
  /Rent/{id}:
    get:
      consumes:
      - application/json
      description: Return rent by id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetRentOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Rent information
      tags:
      - rent
  /Rent/End/{id}:
    post:
      consumes:
      - application/json
      description: Completion of the rent of transport under the rent id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: number
      - in: query
        maximum: 180
        minimum: -180
        name: lat
        required: true
        type: number
      - in: query
        maximum: 180
        minimum: -180
        name: long
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EndRentOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: End rent
      tags:
      - rent
  /Rent/MyHistory:
    get:
      consumes:
      - application/json
      description: Returns the rental history of the current account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetRentsOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Account rent history
      tags:
      - rent
  /Rent/New/{id}:
    post:
      consumes:
      - application/json
      description: Renting the transport for personal use.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: string
      - description: '-'
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.GetRentTransportNewInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetRentTransportNewOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Error'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/handler.Error'
        "412":
          description: Precondition Failed
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Create rent
      tags:
      - rent
  /Rent/Transport:
    get:
      consumes:
      - application/json
      description: Getting transport available for rent by parameters.
      parameters:
      - description: '-'
        in: query
        name: lat
        required: true
        type: number
      - description: '-'
        in: query
        name: long
        required: true
        type: number
      - description: '-'
        in: query
        name: radius
        required: true
        type: number
      - description: '-'
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetRentTransportOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Get transport list
      tags:
      - rent
  /Rent/TransportHistory/{id}:
    get:
      consumes:
      - application/json
      description: Returns the transport rental history.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetRentTransportHistoryOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Transport rent history
      tags:
      - rent
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"